修复类型错误

+++ src/main/kotlin/org/mechdancer/statemachine/core/builder/LinearStateMachineBuilder.kt
@@ -52,7 +52,7 @@ class LinearStateMachineBuilder : IStateMachineBuilder {
 		states.add { engine.run() }
 	}
 
-	fun delay(limit: Double, timeUnit: TimeUnit) = apply {
+	fun delay(limit: Long, timeUnit: TimeUnit) = apply {
 		drive(delayMachine(limit, timeUnit))
 	}
